requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
requests>=2.31
joblib>=1.3
python-dotenv>=1.0
flask>=3.0
pyarrow>=14.0
